Fix initializing workers_nbindid
When bindids are set from the user config, we should just use the number
they gave.

When bindids are set by default, we should just create one series, and
let the binding code iterate over it if needed. Otherwise
duplicate entries in currently_bound/shared would be created for
the same core.