Fix case where data for W-only was prefetched
In that case we didn't mark it initialized during the fetch request, to
avoid dumbly writing back the data if memory gets tight before the task has
a chance to start. But we do want to mark it initialized when the task
is actually about to start. Otherwise we'll be trying to call cl_init.

Fixes #11