Add STARPU_MPI_GPUDIRECT support to HIP
* Tested with HIP-ROCm and HIP-CUDA
* For HIP-CUDA we disable gpudirect if we can't detect the MPI is compiled with CUDA support (like it's done for CUDA)
* For HIP-ROCm however, I propose to enable it at the user request even if we can't detect ROCm support.

The reason is that MPIX_Query_rocm_support isn't available on current release of OpenMPI, but in our case we use UCX to provide ROCm support. While having GPU support in both MPI and UCX is preferred (for collective), we can still have GPU-aware features without support in MPI (afaik). We could also detect ROCm support using ucp_context_query, but this should already be done in MPIX_Query_cuda_support so we can expect it will be the same for ROCm.

Instead I added a warning to inform the user about what starpu is doing.